requirements.txt
PyYAML>=5.4
discord.py>=2.0.0
